:sparkles: feat : 구한 pointToAdd와 프론트에서 보낸 pointToAdd가 같은지 확인하는 기능 구현

+++ src/test/java/cart/domain/ProductTest.java
@@ -12,7 +12,7 @@ class ProductTest {
         final Product product = new Product("사과", 10000, "aa", 10.0, true);
         
         // when
-        final int actualPointToAdd = product.calculatePointToAdd();
+        final long actualPointToAdd = product.calculatePointToAdd();
         
         // then
         assertThat(actualPointToAdd).isEqualTo(1000);
